mini-splatting2/requirements.txt
tqdm
plyfile
open3d
submodules/diff-gaussian-rasterization
submodules/diff-gaussian-rasterization_ms
submodules/simple-knn
submodules/fused-ssim
Pillow>=9.0.0
torchvision